+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.2.8
-- **[fix]** - `ListAvatar` secondary action position
+## v2.3.0
+- **[feat]** - Add `TimeLine` variant "hover"
+- **[feat]** - Add `TimeLine` collapsible feature